requirements.txt
python-telegram-bot>=21.0,<23.0
pyyaml>=6.0
SpeechRecognition>=3.10
